% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_xml.R
\name{build_xml}
\alias{build_xml}
\title{#' Build EndNote xml from MS Forms data}
\usage{
build_xml(record_list, write = FALSE)
}
\arguments{
\item{record_list}{Required. A list object produced by \code{"load_data()"} or \code{"load_example_data()"}.}

\item{write}{Optional boolean. Default = \code{FALSE}. TRUE produces a dialog box for the user to save their xml output. **NOTE** The user must provide the file extension .xml}
}
\value{
\code{xml_output}: xml document of data parsed from \code{"record_list"}.

\code{xml}: Optional, only available if \code{write = TRUE}. A .xml file, saved to the user's computer, that can be imported into EndNote to add records to an EndNote library.
}
\description{
Build xml in \code{\link{xmlGenerator}}
}
\examples{
xmlGenerator::build_xml(record_list) # `write` defaults to FALSE
xmlGenerator::build_xml(record_list, write = TRUE)
# provide a filename in dialog box. E.g., my_xml.xml

}
