% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example of xmlGenerator-acceptable data}
\format{
A data frame with 19 rows and 158 variables.
}
\usage{
example_data
}
\description{
A dataset containing example MS Forms data in the the column name and column order requuired by xmlGenerator.
 The use \code{colnames(example_data)} to extract column names and column order.
}
\keyword{datasets}
