% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ref_type_lookup.R
\name{make_ref_type_lookup}
\alias{make_ref_type_lookup}
\title{Make a lookup table for reference types}
\usage{
make_ref_type_lookup()
}
\arguments{
\item{none}{}
}
\value{
\code{"example_data"}: a data frame containing reference type (e.g., book, newspaper) key-value pairs and column-index information
}
\description{
Make a reference type lookup table in \code{\link{xmlGenerator}}
}
\examples{
# note: xmlGenerator::make_ref_type_lookup() is a sub-function of `load_example_data()` and `load_data()` and isn't meant to be called on its own.
xmlGenerator::make_ref_type_lookup()

}
