% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_example_data.R
\name{load_example_data}
\alias{load_example_data}
\title{Load example data}
\usage{
load_example_data()
}
\arguments{
\item{none}{}
}
\value{
\code{"example_data"}: a data frame containing the column structure that user-provided data must match. A user's MS Forms spreadsheet must have exactly the same column order and names as \code{"example_data"} to generate valid xml.

\code{"forms_spreadsheet"}: a data frame containing user-provided MS Forms data plus one column that \code{"load_example_data()"} left-joined to \code{"example_data"}: \code{"forms_spreadsheet$value"} that contains key values for reference-type lookups called in \code{"build_xml"}.

\code{"record_list"}: a list of \code{"forms_spreadsheet"} data parsed into fields for xml.
}
\description{
Load example data in \code{\link{xmlGenerator}}
}
\examples{
xmlGenerator::load_example_data()

}
